requirements.txt
aiogram==3.8.0
aiohttp==3.9.5
environs==11.0.0
g4f==0.3.2.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
aiosqlite==0.20.0